refactor(controller): 컨트롤러 테스트 위치 변경
